code cleaning and locus_ji query
removing debug cruft from what is now jaccard
new query
